Chapter11/RecommendationService/requirements.txt
pytest
nameko
py2neo
grpcio==1.7.3
